+++ src/containers/reports/index.tsx
@@ -23,7 +23,8 @@ import {
   ReturnButton,
 } from '../../components/themedComponents';
 import { CSVLink } from 'react-csv';
-import { AppError } from '../../auth/axios';
+// import { AppError } from '../../auth/axios';
+import { AppError } from '../../App';
 import ExportDataForm from '../../components/forms/exportDataForm';
 import { Routes } from '../../App';
