+++ src/utils/test/stringFormat.test.ts
@@ -17,7 +17,8 @@ import { getDateString } from '../stringFormat';
 import { shortHand } from '../stringFormat';
 import { SHORT_HAND_NAMES } from '../../assets/content';
 import { Entry, SiteEntryFields } from '../../containers/treePage/ducks/types';
-import { AppError } from '../../auth/axios';
+// import { AppError } from '../../auth/axios';
+import { AppError } from '../../App';
 import { Websites } from '../../constants';
 
 test('getMoneyString tests', () => {
